Правки
Створив файл appendSearchForm.js
