+++ src/scss/hero.scss
@@ -1,80 +1,82 @@
-.title-container{
+.title-container {
   position: absolute;
-    @media screen and (max-width: 768px) {
-      width: 212px;
-      height: 44px;
-      left: 53px;
-      top: 78px;
+  @media screen and (max-width: 768px) {
+    width: 212px;
+    height: 44px;
+    left: 53px;
+    top: 78px;
 
-      margin-bottom: 25px;
+    margin-bottom: 25px;
   }
 
-    @media screen and(min-width: 768px) and (max-width: 1280px) {
-      width: 353px;
-      height: 74px;
-      left: 208px;
-      top: 40px;
+  @media screen and(min-width: 768px) and (max-width: 1280px) {
+    width: 353px;
+    height: 74px;
+    left: 208px;
+    top: 40px;
 
-      margin-bottom: 27px;
+    margin-bottom: 27px;
   }
-  
-    @media screen and(min-width: 1280px) {
-      width: 590px;
-      height: 122px;
-      left: 345px;
-      top: 40px;
-      
-      margin-bottom: 52px;
-  }
-}
 
-.form-container{
-    @media screen and (max-width: 767px) {
-      position: absolute;
-      width: 270px;
-      height: 105px;
-      left: 25px;
-      top: 147px;
+  @media screen and(min-width: 1280px) {
+    width: 590px;
+    height: 122px;
+    left: 345px;
+    top: 40px;
+
+    margin-bottom: 52px;
   }
 }
 
-.input-container{
+.form-container {
+  @media screen and (max-width: 767px) {
+    position: absolute;
+    width: 270px;
+    height: 105px;
+    left: 25px;
+    top: 147px;
+  }
+}
+
+.input-container {
   width: 270px;
   height: 40px;
 
-    @media screen and (max-width: 767px) {
-      margin-bottom: 25px;
+  @media screen and (max-width: 767px) {
+    margin-bottom: 25px;
   }
 
-    @media screen and(min-width: 768px) and (max-width: 1280px) {
-      position: absolute;
-      left: 99px;
-      top: 141px;
-      margin-bottom: 0;
-      margin-right: 30px;
+  @media screen and(min-width: 768px) and (max-width: 1280px) {
+    position: absolute;
+    left: 99px;
+    top: 141px;
+    margin-bottom: 0;
+    margin-right: 30px;
   }
-  
-    @media screen and(min-width: 1280px) {
-      position: absolute;
-      left: 355px;
-      top: 214px;
-      margin-bottom: 0;
-      margin-right: 30px;
+
+  @media screen and(min-width: 1280px) {
+    position: absolute;
+    left: 355px;
+    top: 214px;
+    margin-bottom: 0;
+    margin-right: 30px;
   }
 }
 
-#input-event{
+#input-event {
   width: 270px;
   height: 40px;
+  padding-left: 10px;
 
   background: rgba(255, 255, 255, 0.2);
   -webkit-backdrop-filter: blur(20px);
   backdrop-filter: blur(20px);
-  
+
+  border: none;
   border-radius: 20px;
 }
 
-.input-search-label{
+.input-search-label {
   position: absolute;
   width: 122px;
   height: 20px;
@@ -90,24 +92,24 @@
   color: #939393;
 }
 
-.select{
+.select {
   width: 270px;
   height: 40px;
 
-    @media screen and(min-width: 768px) and (max-width: 1280px) {
-      position: absolute;
-      left: 399px;
-      top: 141px;
+  @media screen and(min-width: 768px) and (max-width: 1280px) {
+    position: absolute;
+    left: 399px;
+    top: 141px;
   }
-  
-    @media screen and(min-width: 1280px) {
-      position: absolute;
-      left: 655px;
-      top: 214px;
+
+  @media screen and(min-width: 1280px) {
+    position: absolute;
+    left: 655px;
+    top: 214px;
   }
 }
 
-#input-country{
+#input-country {
   background: none;
   border: none;
 
@@ -122,7 +124,7 @@
   border-radius: 20px;
 }
 
-.select> select:first-child{
+.select > select:first-child {
   padding-left: 30px;
 
   font-family: Montserrat;
