+++ src/templates/hero.hbs
@@ -10,9 +10,9 @@
         name='event'
         id='input-event'
         class='input-search'
-        placeholder=' '
+        placeholder='Start searching'
       />
-      <label class='input-search-label' for='input-event'>Start searching</label>
+      <label class='input-search-label' for='input-event'></label>
     </div>
 
     <div class='select input-group select-container'>
